pyproject.toml
[tool.poetry]
name = "twitlists"
version = "0.1.0"
description = ""
authors = ["Peyton Walters <pawalt@hey.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
tweepy = "^4.4.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
